Small bug fixed & Class names changed
Fixed this bug from #6, where the delay is now. But the bug can't be fixed if your server is using PlaceholderAPI, because the plugin register the different expansions.
