+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>martinersej</groupId>
     <artifactId>LeuxServerSelector</artifactId>
-    <version>1.0.3</version>
+    <version>1.0.4</version>
     <packaging>jar</packaging>
 
     <name>LeuxServerSelector</name>
