--- src/main/java/martinersej/Utils/Server.java
@@ -1,9 +1,0 @@
-package martinersej.Utils;
-
-import martinersej.LeuxServerSelector;
-
-public class Server {
-    public static String getPrefix() {
-        return LeuxServerSelector.configYML.getString("Prefix");
-    }
-}
