+++ src/main/java/martinersej/utils/Server.java
@@ -1,0 +1,9 @@
+package martinersej.utils;
+
+import martinersej.LeuxServerSelector;
+
+public class Server {
+    public static String getPrefix() {
+        return LeuxServerSelector.configYML.getString("Prefix");
+    }
+}
