+++ src/main/java/martinersej/utils/SkullCreator.java
@@ -1,4 +1,4 @@
-package martinersej.Utils;
+package martinersej.utils;
 
 import com.mojang.authlib.GameProfile;
 import com.mojang.authlib.properties.Property;
