+++ src/main/java/martinersej/utils/UniversalMaterial.java
@@ -1,8 +1,7 @@
-package martinersej.Utils;
+package martinersej.utils;
 
 import martinersej.LeuxServerSelector;
 import org.bukkit.Material;
-import org.bukkit.block.Block;
 import org.bukkit.inventory.ItemStack;
 
 public enum UniversalMaterial{
